requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
supabase==2.3.4
python-dotenv==1.0.1
httpx==0.26.0
numpy==1.26.4
python-multipart==0.0.9
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.4.0
joblib>=1.3.0
pydantic-settings==2.1.0
